# Improve performance by removing `fnv` dependency