--- a/Cargo.toml
+++ b/Cargo.toml
@@ -19,7 +19,6 @@ edition = "2021"
 home = "0.5.0"
 petgraph = "0.6.0"
 nom = "7.0"
-fnv = "1.0"
 memchr = "2.0"
 once_cell = "1.0"
 tree_magic_db = { version = "3.0", path = "./magic_db" , optional = true }