--- a/benches/text/plain
+++ b/benches/text/plain
@@ -1,1 +1,0 @@
-This is just standard text.