--- a/src/fdo_magic/builtin/mod.rs
+++ b/src/fdo_magic/builtin/mod.rs
@@ -2,7 +2,7 @@
 
 use super::MagicRule;
 use crate::Mime;
-use fnv::FnvHashMap;
+use std::collections::HashMap;
 use once_cell::sync::Lazy;
 use petgraph::prelude::*;
 
@@ -13,10 +13,10 @@ pub mod init;
 mod runtime;
 
 /// Preload alias list
-static ALIASES: Lazy<FnvHashMap<Mime, Mime>> = Lazy::new(init::get_aliaslist);
+static ALIASES: Lazy<HashMap<Mime, Mime>> = Lazy::new(init::get_aliaslist);
 
 /// Load magic file before anything else.
-static ALL_RULES: Lazy<FnvHashMap<Mime, DiGraph<MagicRule<'static>, u32>>> = Lazy::new(|| {
+static ALL_RULES: Lazy<HashMap<Mime, DiGraph<MagicRule<'static>, u32>>> = Lazy::new(|| {
     #[cfg(feature = "with-gpl-data")]
     return super::ruleset::from_u8(tree_magic_db::magic()).unwrap_or_default();
     #[cfg(not(feature = "with-gpl-data"))]