… Cargo.lock
… benches/application/x-7z-compressed
… benches/application/zip
… benches/image/gif
… benches/image/png
… src/lib.rs