@aehancoc 2024-04-13
I'm realizing the benchmarks might be easier to read if the benchmarks are ported to [Criterion](https://crates.io/crates/criterion). There's almost certainly a good bit of jitter from the initial lazy load.