@mbrubeck 2024-04-15
> Hey, long time no see. Thanks for maintaining this in my absence. (In case you're wondering, my old account was tied to my university email, which I lost access to.)

You’re welcome!  And if you’re interested at any point in being the maintainer (or a co-maintainter) of this package again, just let me know.  I don’t actually use this library in my other projects anymore, though I’m happy to continue keeping it up to date for others.